Plan for parallelisation of image_to_node
I have a reasonable idea of how to perform the whole image_to_node procedure natively in torch, so as to allow parallelisation.

See lines 147 ff. of model.py for detailed notes.

Summarily, on Monday, complete this work by directly padding the superpixels (formerly sp_tensor) and making necessary adjustments to the subsequent operations, such as view_as_windows, in line with what is required.
Consider also using the .scatter() method to perform the operation converting the superpixel_mask into the superpixels (sp_tensor) as these are effectively the same object; sp_mask is just a sparse version of superpixels, which is one-hot, albeit along an additional dimension