Saving parallelised version of model with notes
This commit serves to show the parallelised code with the relevant checks that demonstrate that it produces identical output to the legacy functions, which utilise the cpu.

In particular, lines 570 ff. and the check on line 583 of modelparallelise.py serve to show that the image_to_nodes function returns identical output whilst not performing all operations natively using Torch.

Previous versions of the functions are saved with the suffix _cpu, Luca's original implementation, _parallel_over_t, the version which uses Torch to vectorise the code over the time dimension and finally the regular image_to_nodes and extract_sp_feat functions, which now only use Torch for production of the superpixel-based node representations.

Note that the following commit removes notes, checks and comments to consolidate the code.
Note Implemention of the variable superpixels over time will likely require an amendment to this vectorised implementation.

AK 28th Oct. 21