Completed parallelised model
This commit contains a neat version of the parallelised image_to_nodes function producing representations from superpixels.

The legacy CPU-based versions of the functions are retained with the appended suffix _cpu.
In particular, these can be checked against the new implementation by inserting the following
at line 350 of modelparallelise.py to check the identity of the representations produced by
the new and old functions:

This check has already been passed, i.e. it returns True indicating identity of the whole tensor
across implementations.

Hopefully the new code will be faster, and at the very least can be easily parallelised.