Save Superpixel Teacher-Student before Training
The training loop has been modified as well as the train.sh script with prob=0 so as to always use the superpixel method.
Note: The train script feeds in an incompatible tensor of image patches if using the patches model, i.e. when prob != 0.

The visualisation script in utils is amended to point to our space on W&B.

All main functionality is in the CRWSuperpixelTeacherStudent class in the teacherstudentsuperpixels.py script.