Safely extract Student for evaluation
I added a function called extract_student_state_dict under the test_utils to allow for extraction of the Student model when evaluating checkpoints from Teacher-Student training that contain both Teacher and Student state_dicts in the same model state_dict.
I also added a test argument allowing you to specify this is what you are evaluating so that the student is extracted when running test.py (via test.sh).