Fixed issue with centroids and superpixels
We identified the issue with the segmentations returned. Some images are very uniform, causing presence of zero-size superpixels.

Solution: We use an epsilon when computing the superpixel representations in image_to_node.
See the EPS added to the mask term when performing the calculation of superpixel representations as the mean of their latent space pixel embeddings