Implement Dilation of Superpixels
This single commit contains all additional work to implement dilation of superpixel masks according to a dilation kernel (shape according to L1 distance, L2 (circle) or a cross kernel, setting post-dilation mask values to 1 when the superpixel is present in the central row or column of the kernel.

Kernel size defaults to 51 (and is forced to be an odd size).

Dilation is set via the dilate-superpixels argument, with kernel shape and size also set via args.

Implementation of the dilation is performed via a depthwise-separable 2D convolution. It might more aptly be named Binary Mask Dilation, as this is a special case where the dilation can be directly performed by convolution followed by a greater than zero comparison over the 2D convolved image